+++ components/ArticleDetail/CommentList.js
@@ -3,8 +3,6 @@ import profileIcon from '@/public/ic_profile.png';
 import noComment from '@/public/no_comment.png';
 import Image from 'next/image';
 import styles from '@/styles/Comment.module.css';
-import { useEffect, useState } from 'react';
-import axios from 'axios';
 
 function CreateDate({ createDate }) {
   const createdDate = new Date(createDate.createdAt);
@@ -17,62 +15,9 @@ function CreateDate({ createDate }) {
   return result;
 }
 
-export default function Comments({ comments, articleId }) {
-  const [comment, setComment] = useState('');
-  const [submit, setSubmit] = useState(false);
-
-  const handleComment = (event) => {
-    setComment(event.target.value);
-  };
-
-  async function postComment() {
-    try {
-      const res = await axios.post(
-        `https://sprint-be-h8kw.onrender.com/comments`,
-        {
-          content: comment,
-          articleId: articleId,
-          userId: '3160c83b-8dcc-4ca2-9d51-717c5246d414',
-        }
-      );
-      console.log(res.data);
-    } catch (error) {
-      console.error('Error posting data:', error);
-    }
-  }
-
-  useEffect(() => {
-    if (comment) {
-      setSubmit(true);
-    } else {
-      setSubmit(false);
-    }
-  }, [comment]);
-
-  function handleSubmit(e) {
-    postComment();
-    setComment('');
-  }
-
+export default function CommentList({ comments }) {
   return (
-    <div className={styles.submit}>
-      <div className={styles.comment}>댓글달기</div>
-      <textarea
-        placeholder='댓글을 입력해 주세요.'
-        type='text'
-        value={comment}
-        onChange={handleComment}
-        className={styles.inputComment}
-      />
-      <button
-        disabled={!submit}
-        className={submit ? styles.submitBtn : styles.btn}
-        type='button'
-        onClick={handleSubmit}
-      >
-        등록
-      </button>
-
+    <>
       {comments.length > 0 ? (
         comments.map((comment) => (
           <div key={comment.id}>
@@ -106,6 +51,6 @@ export default function Comments({ comments, articleId }) {
           priority
         />
       )}
-    </div>
+    </>
   );
 }
