+++ pages/index.js
@@ -1,4 +1,3 @@
-import Link from 'next/link';
 import styles from '@/styles/Home.module.css';
 
 export default function Home() {
